Fix signature of mem_map() to use uint64_t for size
This fixes https://github.com/mcostalba/scoutfish/issues/31

+++ src/misc.cpp
@@ -329,7 +329,7 @@ void bindThisThread(size_t idx) {
 
 } // namespace WinProcGroup
 
-void mem_map(const char* fname, void** baseAddress, uint64_t* mapping, size_t* size) {
+void mem_map(const char* fname, void** baseAddress, uint64_t* mapping, uint64_t* size) {
 
 #ifndef _WIN32
     struct stat statbuf;
